Add unilateral exit support to Nitro Ark
Expose a new `startExitForEntireWallet` API through the Rust FFI and Nitro C++/TypeScript layers, alongside the existing exit sync plumbing.
